apps/worker/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# To update, run:
#
#    pip-compile requirements.in
#
--index-url https://pypi.org/simple
--extra-index-url https://pypi.org/simple

aiofiles==23.2.1
    # via -r requirements.in
aiohttp==3.9.1
    # via -r requirements.in
aioredis==2.0.1
    # via -r requirements.in
anyio==4.1.0
    # via httpx
arrow==1.3.0
    # via -r requirements.in
asyncio==3.4.3
    # via -r requirements.in
black==23.11.0
    # via -r requirements.in
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via -r requirements.in
cryptography==41.0.8
    # via
    #   -r requirements.in
    #   python-jose
fastapi==0.104.1
    # via -r requirements.in
h11==0.14.0
    # via httpx
httpcore==1.0.2
    # via httpx
httpx==0.25.2
    # via -r requirements.in
idna==3.6
    # via anyio
isort==5.12.0
    # via -r requirements.in
jinja2==3.1.2
    # via litellm
jose==3.3.0
    # via python-jose
litellm==1.13.4
    # via -r requirements.in
markupsafe==2.1.3
    # via jinja2
mypy==1.7.1
    # via -r requirements.in
openai==1.3.7
    # via -r requirements.in
opentelemetry-api==1.21.0
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.42b0
    # via -r requirements.in
opentelemetry-sdk==1.21.0
    # via opentelemetry-instrumentation-fastapi
orjson==3.9.10
    # via -r requirements.in
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pgvector==0.2.4
    # via -r requirements.in
pre-commit==3.6.0
    # via -r requirements.in
psycopg[binary]==3.1.12
    # via -r requirements.in
pydantic==2.5.0
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic[email]==2.5.0
    # via -r requirements.in
pydantic-settings==2.1.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
redis==5.0.1
    # via aioredis
requests==2.31.0
    # via litellm
rich==13.7.0
    # via -r requirements.in
ruff==0.1.8
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
slowapi==0.1.9
    # via -r requirements.in
sse-starlette==1.8.0
    # via -r requirements.in
structlog==23.2.0
    # via -r requirements.in
tenacity==8.2.3
    # via -r requirements.in
typer==0.9.0
    # via -r requirements.in
typing-extensions==4.8.0
    # via
    #   anyio
    #   pydantic
    #   pydantic-core
urllib3==2.1.0
    # via requests
uvicorn[standard]==0.24.0
    # via -r requirements.in
watchdog==3.0.0
    # via -r requirements.in